medical_trancriber/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
python-dotenv>=0.19.0
pydantic>=1.8.0
httpx<0.28
deepgram-sdk==3.10.0
google-cloud-aiplatform>=1.38.1
google-cloud-secret-manager>=2.23.1